#2 #add #sender - send phase feature
- add MessageRaw
- add TelegramBotMessagePreparer, and impl AssistantTelegramBotMessagePreparer
- add DocumentMedia
- replace SendText to Message
- add VisualMedia
- impl AssistantTelegramBotMessageSender
- add TelegramBotDocumentMediaSender
- add TelegramBotVisualMediaSender
- add TelegramBotUpdateProcessor, and impl AssistantTelegramBotUpdateProcessor

+++ src/main/kotlin/org/gnori/bgauassistantbot/common/phase/entity/PhaseEntity.kt
@@ -9,6 +9,6 @@ class PhaseEntity(
     @Id val id: UUID,
     val name: String,
     val description: String,
-    val shortId: Long,
+    val shortId: Int,
     val parentId: UUID?
 )
